+++ api/__init__.py
@@ -2,3 +2,4 @@ from api.api_attractions import *
 from api.api_booking import *
 from api.api_users import *
 from api.api_order import *
+from api.db import *
