+++ api/api_attractions/__init__.py
@@ -1,0 +1,1 @@
+from api.api_attractions.api_attractions import *
