+++ api/api_booking/__init__.py
@@ -1,0 +1,1 @@
+from api.api_booking.api_booking import *
